requirements.txt
alembic==1.7.5
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.2
python-dotenv==0.19.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
